✨ feat: Add customer registration functionality
Adds a new POST endpoint to register a new customer. The submitted
customer data is validated for uniqueness of email, and then persisted
to the database. The customer service has been updated to handle the new
functionality, and a new CustomerRegistrationRequest class has been
added to encapsulate the customer data.

0084_Testing_POST_Request